requirements.txt

Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
pytz==2022.7
sqlparse==0.4.3
